chore: Update install.sh to conditionally execute docker compose commands based on the presence of docker-compose.yml

+++ install.sh
@@ -7,7 +7,10 @@ source .env
 mkdir -p /opt/backup-manager
 cd /opt/backup-manager
 
-VERSION=$VERSION docker compose down --volumes --remove-orphans --rmi all
+if [ -f docker-compose.yml ]; then
+    VERSION=$VERSION docker compose down --volumes --remove-orphans --rmi all
+fi
+
 
 curl -o docker-compose.yml https://raw.githubusercontent.com/brayandm/backup-manager/$VERSION/docker-compose.yml
 VERSION=$VERSION docker compose up -d
